Move freeze_args to ocrd.utils.introspect
