+++ ocrd/requirements.txt
@@ -11,4 +11,3 @@ Deprecated == 1.2.0
 memory-profiler >= 0.58.0
 sparklines >= 0.4.2
 python-magic
-frozendict>=2.3.4
