+++ ocrd_utils/requirements.txt
@@ -5,3 +5,4 @@ numpy
 atomicwrites >= 1.3.0
 importlib_metadata;python_version<'3.8'
 importlib_resources;python_version<'3.8'
+frozendict>=2.3.4
